Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-cross-in-same-self-block-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8080/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=%27self%27 because it does not appear in the frame-ancestors directive of the Content Security Policy.
A 'frame-ancestors' CSP directive with a value 'same' should block render in same-origin nested frames. This test PASSED if you do not see any content in the inner-most nested iframe below. Otherwise, it FAILED.



--------
Frame: '<!--frame1-->'
--------
Testing a cross-origin child with a policy of "'self'" nested in a same-origin parent.



--------
Frame: '<!--frame2-->'
--------
